+++ src/engine/math/Vector.h
@@ -11,13 +11,15 @@
 template <class type, int size>
 class Vector {
 protected:
-    const int length = size;
+//    const int length = size;
     type array[size];
 //    const boost::array<type, size> array{0};
 //    const std::array<type, size> arr;
 public:
     Vector();
     Vector(std::vector<type> v);
+//    Vector(const Vector &vector);
+
     type& operator[](int i);
     Vector& normalize();
     type dot(const Vector &other);
@@ -35,169 +37,6 @@ public:
     Vector operator*(const Vector& other) const;
     Vector operator*(type t) const;
 };
-
-template<class type, int size>
-Vector<type, size>::Vector(std::vector<type> v) {
-    for (int i = 0; i < v.size(); ++i) {
-        this->array[i] = v[i];
-    }
-}
-
-template<class type, int size>
-Vector<type, size>::Vector() {
-    for (int i = 0; i < size; ++i) {
-        this->array[i] = 0;
-    }
-}
-
-template<class type, int size>
-type& Vector<type, size>::operator[](int i) {
-    if (i >= 0 && i < size) {
-        return this->array[i];
-    } else {
-        throw std::out_of_range("Parameter must be integer value between 0 and 2");
-    }
-}
-
-template<class type, int size>
-Vector<type, size>& Vector<type, size>::normalize() {
-    double d = 0;
-    auto square = [](type i) { return i*i; };
-    for (auto e: this->array) {
-        d += square(e);
-    }
-
-    d = sqrt(d);
-
-    for(int i = 0; i < size; i++) {
-        this->array[i] /= d;
-    }
-
-    return *this;
-}
-
-template<class type, int size>
-type Vector<type, size>::dot(const Vector &other) {
-    type result = 0;
-
-    for (int i = 0; i < size; ++i) {
-        result += this->array[i] * other.array[i];
-    }
-
-    return result;
-}
-
-template<class type, int size>
-Vector<type, size>& Vector<type, size>::operator=(const Vector &other) {
-    if (this != &other) {
-        for(int i = 0; i < size; i++) {
-            this->array[i] = other.array[i];
-        }
-    }
-    return *this;
-}
-
-template<class type, int size>
-Vector<type, size>& Vector<type, size>::operator+=(const Vector &other) {
-    for(int i = 0; i < size; i++) {
-        this->array[i] += other.array[i];
-    }
-    return *this;
-}
-
-template<class type, int size>
-Vector<type, size>& Vector<type, size>::operator+=(type t) {
-    for(int i = 0; i < size; i++) {
-        this->array[i] += t;
-    }
-    return *this;
-}
-
-template<class type, int size>
-Vector<type, size> Vector<type, size>::operator+(const Vector &other) const {
-    Vector<type, size> result;
-    for (int i = 0; i < size; ++i) {
-        result.array[i] = this->array[i] + other.array[i];
-    }
-    return result;
-};
-
-template<class type, int size>
-Vector<type, size> Vector<type, size>::operator+(type t) const {
-    Vector<type, size> result;
-    for (int i = 0; i < size; ++i) {
-        result.array[i] = this->array[i] + t;
-    }
-    return result;
-};
-
-template<class type, int size>
-Vector<type, size>& Vector<type, size>::operator-=(const Vector &other) {
-    for(int i = 0; i < size; i++) {
-        this->array[i] -= other.array[i];
-    }
-    return *this;
-}
-
-template<class type, int size>
-Vector<type, size>& Vector<type, size>::operator-=(type t) {
-    for(int i = 0; i < size; i++) {
-        this->array[i] -= t;
-    }
-    return *this;
-}
-
-template<class type, int size>
-Vector<type, size> Vector<type, size>::operator-(const Vector &other) const {
-    Vector<type, size> result;
-    for (int i = 0; i < size; ++i) {
-        result.array[i] = this->array[i] - other.array[i];
-    }
-    return result;
-}
-
-template<class type, int size>
-Vector<type, size> Vector<type, size>::operator-(type t) const {
-    Vector<type, size> result;
-    for (int i = 0; i < size; ++i) {
-        result.array[i] = this->array[i] - t;
-    }
-    return result;
-}
-
-template<class type, int size>
-Vector<type, size>& Vector<type, size>::operator*=(const Vector &other) {
-    for(int i = 0; i < size; i++) {
-        this->array[i] *= other.array[i];
-    }
-    return *this;
-}
-
-template<class type, int size>
-Vector<type, size>& Vector<type, size>::operator*=(type t) {
-    for(int i = 0; i < size; i++) {
-        this->array[i] *= t;
-    }
-    return *this;
-}
-
-template<class type, int size>
-Vector<type, size> Vector<type, size>::operator*(const Vector &other) const {
-    Vector<type, size> result;
-    for (int i = 0; i < size; ++i) {
-        result.array[i] = this->array[i] * other.array[i];
-    }
-    return result;
-}
-
-template<class type, int size>
-Vector<type, size> Vector<type, size>::operator*(type t) const {
-    Vector<type, size> result;
-    for (int i = 0; i < size; ++i) {
-        result.array[i] = this->array[i] * t;
-    }
-    return result;
-}
 
 // ------------------ Utils ---------------------
